Move editor.utils.code_analysis to editor.lsp
