+++ spyder/plugins/editor/lsp/decorators.py
