+++ spyder/plugins/editor/lsp/transport/consumer.py
