+++ spyder/plugins/editor/lsp/transport/producer.py
