+++ spyder/plugins/editor/plugin.py
@@ -53,7 +53,7 @@ from spyder.api.preferences import PluginConfigPage
 from spyder.preferences.runconfig import (ALWAYS_OPEN_FIRST_RUN_OPTION,
                                       get_run_configuration,
                                       RunConfigDialog, RunConfigOneDialog)
-from spyder.plugins.editor.utils.code_analysis import (
+from spyder.plugins.editor.lsp import (
     LSPRequestTypes, LSPEventTypes, TextDocumentSyncKind)
