+++ spyder/plugins/editor/widgets/tests/test_warnings.py
@@ -18,7 +18,7 @@ from spyder.plugins.editor.widgets.codeeditor import CodeEditor
 from spyder.py3compat import to_binary_string
 from spyder.utils.codeanalysis import check_with_pyflakes, check_with_pep8
 from spyder.plugins.lspmanager import LSPManager
-from spyder.plugins.editor.utils.code_analysis import LSPEventTypes
+from spyder.plugins.editor.lsp import LSPEventTypes
 from spyder.py3compat import PY2
